IGAPP-710: Move forceNewLineAfterChar.ts to right dir
