+++ native/src/components/Header.tsx
@@ -14,7 +14,7 @@ import dimensions from '../constants/dimensions'
 import { StoreActionType } from '../redux/StoreActionType'
 import { DISCLAIMER_ROUTE, SEARCH_ROUTE, SETTINGS_ROUTE } from 'api-client/src/routes'
 import navigateToLanding from '../navigation/navigateToLanding'
-import { forceNewlineAfterChar } from '../services/forceNewLineAfterChar'
+import { forceNewlineAfterChar } from '../utils/forceNewLineAfterChar'
 import sendTrackingSignal from '../utils/sendTrackingSignal'
 
 const Horizontal = styled.View`
