+++ native/src/utils/forceNewLineAfterChar.ts
